Add files via upload
The program is to run a quiz where in we have the questions in the Questions.py file. The questions are of True/False type. For each correct answer the user gets +2 points and for each incorrect answer the user gets -1 point. For running the program, we have to make sure that all the files are in the same directory and run the main.py file to get the desired output(quiz).